+++ NEWS.md
@@ -1,4 +1,4 @@
-# brms.mmrm 0.0.2.9003 (development)
+# brms.mmrm 0.1.0
 
 * Deprecate `brm_simulate()` in favor of `brm_simulate_simple()` (#3). The latter has a more specific name to disambiguate it from other simulation functions, and its parameterization conforms to the one in the methods vignette.
 * Add new functions for nuanced simulations: `brm_simulate_outline()`, `brm_simulate_continuous()`, `brm_simulate_categorical()` (#3).